refactor: ユースケースのディレクリ整理
